FastAPI/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
astroid==3.2.4
black==24.8.0
cffi==1.17.1
click==8.1.7
cryptography==43.0.3
dill==0.3.8
distlib==0.3.9
ecdsa==0.19.0
fastapi==0.112.1
filelock==3.16.1
h11==0.14.0
idna==3.7
isort==5.13.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.1
passlib==1.7.4
pathspec==0.12.1
pbr==6.1.0
peewee==3.17.6
platformdirs==4.3.3
pyasn1==0.6.1
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.9.0
pylint==3.2.7
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.12
rsa==4.9
six==1.16.0
sniffio==1.3.1
starlette==0.38.2
stevedore==5.3.0
tomlkit==0.13.2
typing_extensions==4.12.2
uvicorn==0.30.6
